requirements.txt
requests>=2.31.0
pandas>=2.1.0
streamlit>=1.32.0
plotly>=5.18.0
pyyaml>=6.0.1
click>=8.1.7
rich>=13.7.0
python-dotenv>=1.0.1
tabulate>=0.9.0
google-cloud-storage>=2.14.0
